Squashed commit of the following:
commit 86b59b13decb1125acece350594a585613050be0
Author: Toby Sharp <toby@hornetnode.com>
Date:   Mon Jul 14 22:25:36 2025 +0000

    BlockSync progress.

commit 1e5fba94ac2b0bda6da29a922d61f0a69ce7e37d
Author: Toby Sharp <toby@hornetnode.com>
Date:   Mon Jul 14 22:16:55 2025 +0000

    Improve NoHeadersSyncManager for block validation unit test.

commit 3ae77b1fffaaa1b6ff709c36ff529bd21fb24ac1
Author: Toby Sharp <toby@hornetnode.com>
Date:   Mon Jul 14 22:15:37 2025 +0000

    Add HeaderTimechain::Find taking both height and hash to uniquely and quickly specify a header.

commit 3a6d07e70edb19a31fb1305401d33d66bbfee08f
Author: Toby Sharp <toby@hornetnode.com>
Date:   Mon Jul 14 22:14:38 2025 +0000

    Drop peer when remote connection closed.

commit ce20f4206176d5b903685ae692f079fbe9786d12
Author: Toby Sharp <toby@hornetnode.com>
Date:   Sat Jul 12 22:34:49 2025 +0000

    Fix peer disconnection cleanup in ProtocolLoop, and allow for the case of zero connected peers.

commit d3b4e0a06fa72bb1c955e872c5c1b6f44c583aaa
Author: Toby Sharp <toby@hornetnode.com>
Date:   Sat Jul 12 22:33:33 2025 +0000

    Share the protocol block among multiple owners.

commit a61b1c1a69d3d3aa95daaf779cd0d03c2b7cc9ec
Author: Toby Sharp <toby@hornetnode.com>
Date:   Sat Jul 12 22:31:40 2025 +0000

    Add protocol::Block::Genesis and protocol::Block::SizeBytes and unit tests.

commit 65c696f63467d0374f425396270f2bc65185d71b
Author: Toby Sharp <toby@hornetnode.com>
Date:   Sat Jul 12 22:29:22 2025 +0000

    Add string literal suffix _bytes.

commit 17fd41a2f7b655fbca920e6986e7553664f2d026
Author: Toby Sharp <toby@hornetnode.com>
Date:   Fri Jul 11 08:42:19 2025 +0000

    Add unit tests for constructing the genesis block.

commit 7d78f89513e940eb6a7ff8ed1cd9053117fb323b
Author: Toby Sharp <toby@hornetnode.com>
Date:   Thu Jul 10 16:58:05 2025 +0000

    Swtich size_t to int.

commit 2affb37466192fbc61dabae5c2a09d451f8377d5
Author: Toby Sharp <toby@hornetnode.com>
Date:   Thu Jul 10 04:36:48 2025 +0000

    Improvements in ProtocolLoop. Changed polling to be once per loop for both reads and writes. Added a max processing time per frame. Randomized peer order for read/write to avoid structural bias.

commit a92a3d3c3723148926b957328ba9e263e30adcc7
Author: Toby Sharp <toby@hornetnode.com>
Date:   Wed Jul 9 04:09:04 2025 +0000

    Make PeerRegistry mutaters private and accessible only to PeerManager.

commit 35f690eef0e89c087c364247c2d965511f4506fc
Author: Toby Sharp <toby@hornetnode.com>
Date:   Wed Jul 9 03:58:08 2025 +0000

    Changed EventHandler to use SharedPeer rather than WeakPeer in handler methods. Added OnHandshakeComplete event.

commit fe82bee6c35db554d9c7cae4fa19eabc6d074bc0
Author: Toby Sharp <toby@hornetnode.com>
Date:   Tue Jul 8 22:44:29 2025 +0000

    Major refactor for peer ids, impacting the whole hierarchy. Improved separation of concerns. Some tidy-up remaining. All unit tests pass.

commit e87de16d96c144a58865387fb0a27a9a9d8ad567
Author: Toby Sharp <toby@hornetnode.com>
Date:   Thu Jul 3 20:44:43 2025 +0000

    Factor out SyncHander from HeaderSync, and add BlockSync to SyncManager.

commit c74d07b5a98b090473a662d90bede69c2a3ac34e
Author: Toby Sharp <toby@hornetnode.com>
Date:   Wed Jul 2 18:36:35 2025 +0000

    Take transaction count out of BlockHeader, and fix up serialization.

commit 6d430d8bdf531fd24e23bfbed0a53af28fa02384
Author: Toby Sharp <toby@hornetnode.com>
Date:   Wed Jul 2 18:35:26 2025 +0000

    Add GetData and Blok messages.

commit 3198726d9829d93ecf54874a56bd2e194466ddda
Author: Toby Sharp <toby@hornetnode.com>
Date:   Wed Jul 2 18:34:41 2025 +0000

    Add method TransferEnumLE4.

commit d5d9850ca1d50ea2ac561e3aa7ecbbf5cf21c479
Author: Toby Sharp <toby@hornetnode.com>
Date:   Wed Jul 2 18:32:04 2025 +0000

    Add check to ReadVarInt inline with Core behavior.

commit 9b4bf3a4c7528a2d46ab572c2a9426d410e3205d
Author: Toby Sharp <toby@hornetnode.com>
Date:   Wed Jul 2 18:19:35 2025 +0000

    Factor out SubArray into util. Minor stylistic improvements to transaction classes, and added comments.

commit a176d4d6e714897a88c91e46852be0116a7d2fbf
Author: Toby Sharp <toby@hornetnode.com>
Date:   Wed Jul 2 07:26:38 2025 +0000

    Transaction classes and block deserialization.

commit e818ffa5f5abc803e42406b6b3a82918fd138a39
Author: Toby Sharp <toby@hornetnode.com>
Date:   Sun Jun 29 19:02:35 2025 +0000

    Rename HeaderSync::RegisterPeer to StartSync.

commit 79e224b03d7137cb5aa34cd73a373250ca665f99
Author: Toby Sharp <toby@hornetnode.com>
Date:   Sun Jun 29 08:02:38 2025 +0000

    Move HeaderSync from data to node.

commit f063849f7c2baf810c1cfd1b691252109d55c922
Author: Toby Sharp <toby@hornetnode.com>
Date:   Sun Jun 29 07:51:12 2025 +0000

    Fix ValidationView tip advance. Add OnComplete callback for HeaderSync. Add HeaderSyncHandler interface.

+++ src/hornetlib/net/connection.h
@@ -105,10 +105,10 @@ class Connection {
     return write_queue_.size();
   }
 
-  size_t ContinueWrite() {
+  int ContinueWrite() {
     if (!sock_.IsOpen()) return 0;
     const bool is_blocking = sock_.IsBlocking();
-    size_t bytes_written = 0;
+    int bytes_written = 0;
     do {
       while (!write_queue_.empty() && write_queue_.front()->empty())
         write_queue_.pop_front();
